[GA-245] Add ProductDisplayModel.kt

+++ app/src/main/java/com/gthr/gthrcollect/model/network/firebaserealtimedb/FunkoModel.kt
@@ -14,7 +14,7 @@ data class FunkoModel(
     @SerialName(FirebaseRealtimeDatabase.FUNKO_TYPE)
     val funkoType: String,
     @SerialName(FirebaseRealtimeDatabase.HIGHEST_BID_COST)
-    val highestBidCost: Int,
+    val highestBidCost: Double,
     @SerialName(FirebaseRealtimeDatabase.HIGHEST_BID_ID)
     val highestBidID: String,
     @SerialName(FirebaseRealtimeDatabase.IMAGE_URL)
@@ -24,7 +24,7 @@ data class FunkoModel(
     @SerialName(FirebaseRealtimeDatabase.LICENSE)
     val license: String,
     @SerialName(FirebaseRealtimeDatabase.LOWEST_ASK_COST)
-    val lowestAskCost: Int,
+    val lowestAskCost: Double,
     @SerialName(FirebaseRealtimeDatabase.LOWEST_ASK_ID)
     val lowestAskID: String,
     @SerialName(FirebaseRealtimeDatabase.MARKET_VALUE)
